[Bug] Changed radius from Long to Double in GeoLocationConfig

+++ workjam/src/main/kotlin/com/tfowl/workjam/client/model/Employers.kt
@@ -83,7 +83,7 @@ data class Country(
 data class GeolocationConfig(
     val latitude: Double,
     val longitude: Double,
-    val radius: Long
+    val radius: Double
 )
 
 @Serializable
